Cleanup database connection
Removed database name environment variable and replaced with
VeilingDb. This was combined with a migration to ensure the
creation of the database. Furthermore hardcoded database server
value was replaced with a value inside the docker-compose.yml.
Last but not least a skeleton of a database seeder was setup.
All of this is to make the implementation of a CI pipeline easier.