Add simple reusable buttons
Buttons exist in two forms. There are regular simple buttons and
buttons with arrow SVGs. The latter can be expanded later to allow
for any icon to be set. The two buttons also have an "appearance"
property that allows you to define whether they have the full color
primary appearance, or the bordered secondary appearance.

The button size can also change depending on the text size.
The buttons can be used anywhere in the website and should in most
cases be the standard button form.