Create handeling for Verkoper page
The VerkoperDashboard now recieves the data that is being send up from
the components.
The VerkoperDashboard checks if all the fields value's contain the right
datatype and are all filled in.
The logic for fetching goes as follow's
first fetch uploads the image, and recives the URL of the location for
this image, then the JSON is send with the url of said image