Add bidding to backend
Added a new table to the database. For some reason it would not
create the migration properly so I had to reset again. KavelVerkoop
represents the period in which a person can bid for a specific
auction item. As many will be added per auction item as needed
until all containers are sold.

This commit is very much incomplete, as it is lacking the container
selection purchase popup. Bidding isn't bound to a specific user
yet.