Fix server-side bidding.
Server-side bidding works now. There were several problems that
prevented it from functioning earlier, including some unmapped
properties in the Model and some incorrect handling of DateTimes.

The issue of synchronization remains. Server-side price calculation
tends to diverge heavily from client-side. I hoped to fix this by
sending the moment at which the auction starts to the server, and
then synchronizing that across the clients, but it seems this was
not sufficient to prevent desync.